+++ app/policies/annotation_category_policy.rb
@@ -3,6 +3,6 @@ class AnnotationCategoryPolicy < ApplicationPolicy
   default_rule :manage?
 
   def manage?
-    allowed_to?(:manage_assessments?, with: GraderPermissionPolicy)
+    check?(:manage_assessments?, user)
   end
 end