+++ app/policies/criterion_policy.rb
@@ -1,0 +1,8 @@
+# Criterion policy class
+class CriterionPolicy < ApplicationPolicy
+  default_rule :manage?
+
+  def manage?
+    check?(:manage_assessments?, user)
+  end
+end