+++ app/policies/image_annotation_policy.rb
@@ -1,0 +1,3 @@
+# Image annotation policy class
+class ImageAnnotationPolicy < AnnotationPolicy
+end