+++ app/policies/note_policy.rb
@@ -1,3 +1,4 @@
+# Note policy class
 class NotePolicy < ApplicationPolicy
   default_rule :manage?
   alias_rule :edit?, :update?, to: :modify?