+++ app/policies/section_policy.rb
@@ -1,0 +1,8 @@
+# Section policy class
+class SectionPolicy < ApplicationPolicy
+  default_rule :manage?
+
+  def manage?
+    user.admin?
+  end
+end