+++ Changelog.md
@@ -23,6 +23,7 @@
 - Add counts of all/active/inactive students to students table (#4942)
 - Allow feedback files to be updated by uploading a binary file object through the API (#4964)
 - Fix a bug where some error messages reported by the API caused a json formatting error (#4964)
+- Updated all authorization to use ActionPolicy (#4865)
 
 ## [v1.10.3]
 - Allow for more concurrent access to git repositories (#4895)